Add files via upload
Hi everyone,
This is my first Project as a Data Analyst. I am currently looking for employment to make Data Analysis my career. In this project, we utilized Power BI to conduct an in-depth analysis of the World Happiness Report data from 2015 to 2017. Through interactive visualizations and insightful dashboards, we explored and examined the happiness rankings and scores of countries and regions. By diving deep into the data, we were able to identify correlations between happiness and various socio-economic factors. Our findings shed light on the key determinants of well-being and provided valuable insights into the factors that contribute to happiness on a global scale.